Add files via upload
this function works like a ticket office and the goal is to sell tickets to all the people in line. the people only have a single 25, 50 or 100 dollar bill and you have to help them in the sequence that they appear. your till starts out with no money and if you can't give a customer change the function returns "NO" and "YES" if all the customers got a ticket to see a movie. the customers are represented by their dollar note in a list.